requirements/oknesset.txt
# The very basic requirements, hopefully without devel, scraping and video
# download dependencies, others will build upon it
#
# Left out for now from setup
#
# mimeparse - Looks like a requirement by tastypie, will be auto installed
# mimms - Video download
# poppler - required only for simple/management/parse_government_bill,
# kikar-hamedina - leftout: Django==1.5.5 (We work on 1.6.1)
# Also: ipython, south, python-dateutil, django-tastypie  - appear in our base.txt requirement, gunicorn - in vps.txt

PIL
django-pagination==1.0.7
django-tagging==0.3.1
django_compressor==1.2
BeautifulSoup
vobject
feedparser
gdata
django-social-auth>=0.7.19
oauth==1.0.1
python-memcached==1.48
-e git://github.com/openid/python-openid.git#egg=python-openid
django-ratings
-e git://github.com/thornomad/django-hitcount.git#egg=django-hitcount
-e git://github.com/hasadna/django-activity-stream.git@c56a07921b8c796fce35bda6feab4610fdf64e5d#egg=django-activity-stream
-e git://github.com/hasadna/django-planet.git@7a66a954b231391f2ac76e97f2a469f69fd29037#egg=django-planet
-e git://github.com/hasadna/django-annotatetext.git#egg=django-annotatetext
-e git://github.com/pinax/django-mailer.git#egg=django-mailer
-e git://github.com/hasadna/django-avatar.git@e6ed23b44764d75d1f3955472f5afc2c4ddf7bcc#egg=django-avatar
-e git://github.com/hasadna/django-backlinks.git@a90c5db5b4794965b7559ca4631bef8984957a9b#egg=django-backlinks
-e git://github.com/spreeker/django-voting2.git#egg=django-voting2
-e git://github.com/hasadna/django-piston.git@1b51686a5b2ff547429d0b224319d521170f63e2#egg=django-piston
django-debug-toolbar>=1.0
nose
django-nose
django-devserver==0.7.0
django-tinymce==1.5.1b4
django-crispy-forms==1.2.3
django-storages==1.1.8
boto==2.20.1
